Fix crash in ThreadTracks when zooming-in a lot
We sometimes have an issue when zooming-in a lot. Orbit hangs because
falls in an infinite loop while iterating through pixels in
ThreadTracks.

After investigating a bit, I found that the issue is a regression from
https://github.com/google/orbit/pull/4296. We are using 32-bits integer
to get the pixel number, but the requested pixel (end_time of the
current timer) could be very far away and as it overflows, Orbit ends
iterating in an infinite loop.

So, in this PR I'm fixing this issue and updating the unit-tests since
they should have caught the issue.

Test: New unit test fails with the old-code. Now they are green. Load a
capture, see that now it doesn't hang anymore.

+++ src/ClientData/FastRenderingUtils.cpp
@@ -6,7 +6,7 @@
 
 namespace orbit_client_data {
 
-uint32_t GetPixelNumber(uint64_t timestamp_ns, uint32_t resolution, uint64_t start_ns,
+uint64_t GetPixelNumber(uint64_t timestamp_ns, uint32_t resolution, uint64_t start_ns,
                         uint64_t end_ns) {
   uint64_t current_ns_from_start = timestamp_ns - start_ns;
   uint64_t total_ns = end_ns - start_ns;
